+++ core/dbt/artifacts/resources/v1/compiled.py
@@ -1,11 +1,13 @@
 from dataclasses import dataclass, field
 from dbt.artifacts.resources.v1.components import (
+    ParsedNode,
     CompiledNode,
     NodeVersion,
     DeferRelation,
+    MacroDependsOn,
 )
 from dbt_common.contracts.constraints import ModelLevelConstraint
-from dbt.artifacts.resources.v1.config import ModelConfig
+from dbt.artifacts.resources.v1.config import ModelConfig, SeedConfig
 from typing import Literal, Optional, List
 from dbt.artifacts.resources.types import NodeType, AccessType
 from datetime import datetime
@@ -31,4 +33,20 @@ class ModelNode(CompiledNode):
     version: Optional[NodeVersion] = None
     latest_version: Optional[NodeVersion] = None
     deprecation_date: Optional[datetime] = None
+    defer_relation: Optional[DeferRelation] = None
+
+
+@dataclass
+class SqlNode(CompiledNode):
+    resource_type: Literal[NodeType.SqlOperation]
+
+
+@dataclass
+class SeedNode(ParsedNode):  # No SQLDefaults!
+    resource_type: Literal[NodeType.Seed]
+    config: SeedConfig = field(default_factory=SeedConfig)
+    # seeds need the root_path because the contents are not loaded initially
+    # and we need the root_path to load the seed later
+    root_path: Optional[str] = None
+    depends_on: MacroDependsOn = field(default_factory=MacroDependsOn)
     defer_relation: Optional[DeferRelation] = None
